+++ client/src/components/entry-point/header.js
@@ -1,26 +1,26 @@
 import React from 'react';
-import { Box, Link, Text } from 'src/components/style';
+import { Container, Link, Text } from 'src/components/style';
 import { colors } from 'src/models/style';
 import { css } from 'styled-components';
 
 export default function Header() {
   return (
-    <Box flexDirection="row" overrides={styles.container}>
-      <Box flex={0}>
+    <Container flexDirection="row" overrides={styles.container}>
+      <Container flex={0}>
         <Link overrides={styles.link} to="/">
           <Text color={colors.rackleyBlue()} size={20} weight="bold">SomeSchedules</Text>
         </Link>
-      </Box>
-      <Box flex={1} />
-      <Box flex="none" flexDirection="row">
+      </Container>
+      <Container flex={1} />
+      <Container flex="none" flexDirection="row">
         <Link overrides={styles.link} to="/sign-up">
           <Text color={colors.eerieBlack()}>Sign Up</Text>
         </Link>
         <Link overrides={styles.link} to="/sign-in">
           <Text color={colors.eerieBlack()}>Sign In</Text>
         </Link>
-      </Box>
-    </Box>
+      </Container>
+    </Container>
   );
 }
