+++ client/src/components/entry-point/index.js
@@ -4,7 +4,7 @@ import React from 'react';
 import Schedule from 'src/apps/schedule';
 import SignIn from 'src/apps/sign-in';
 import SignUp from 'src/apps/sign-up';
-import { Box } from 'src/components/style';
+import { Container } from 'src/components/style';
 import { Route, Switch } from 'react-router-dom';
 import { colors } from 'src/models/style';
 import { injectGlobal } from 'styled-components';
@@ -31,7 +31,7 @@ injectGlobal`
 
 export default function EntryPoint() {
   return (
-    <Box>
+    <Container>
       <Header />
       <Switch>
         <Route component={SignIn} path="/sign-in" />
@@ -39,6 +39,6 @@ export default function EntryPoint() {
         <Route component={Schedule} path="/schedule" />
         <Route component={Marketing} path="/" />
       </Switch>
-    </Box>
+    </Container>
   );
 }
