+++ client/src/components/style/index.js
@@ -1,3 +1,3 @@
-export { default as Box } from 'src/components/style/box';
+export { default as Container } from 'src/components/style/container';
 export { default as Link } from 'src/components/style/link';
 export { default as Text } from 'src/components/style/text';
