+++ client/src/index.js
@@ -14,12 +14,12 @@ const tree = (
 );
 
 if (process.env.NODE_ENV === 'development') {
-  const RedBox = require('redbox-react').default;
+  const RedContainer = require('redbox-react').default;
 
   try {
     render(tree, root);
   } catch (error) {
-    render(<RedBox error={error} />, root);
+    render(<RedContainer error={error} />, root);
   }
 } else {
   render(tree, root);
